requirements.txt
asgiref==3.4.1
braintree==4.12.0
certifi==2021.5.30
charset-normalizer==2.0.6
crispy-forms-gds==0.2.2
Django==3.2.7
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-mptt==0.13.3
djangorestframework==3.12.4
idna==3.2
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.7
